requirements.txt
streamlit
pandas
numpy
python-dotenv
biopython
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
plotly>=5.0.0
